refactor: unify struct-vs-trait stubs across crates (Phase 15c)
- SongDatabaseAccessor: define trait in beatoraja-types (with Send bound),
  implement in beatoraja-song, replace struct stubs in select/external.
  Move FolderData to beatoraja-types.
- IRConnection: replace struct stubs with real trait from beatoraja-ir
  using Box<dyn IRConnection> in select, Arc<dyn IRConnection> in result.
  Replace LeaderboardEntry, LR2IRConnection, LR2GhostData, IRScoreData,
  IRChartData, IRPlayerData, IRResponse, IRTableData stubs with imports.
- BMSPlayerInputProcessor: unify analog method params (i32 → usize).
- TableData/TableFolder: add getters to real types in beatoraja-core;
  stubs kept due to CourseData field cascade (deferred).

+++ lr2oraja-rust/crates/beatoraja-select/src/music_select_key_property.rs
@@ -75,8 +75,8 @@ impl MusicSelectKeyProperty {
         let mut d_ticks = 0;
         for i in 0..assign.len() {
             for &index in assign[i] {
-                if code == index && input.is_analog_input(i as i32) {
-                    d_ticks += input.get_analog_diff_and_reset(i as i32, 200);
+                if code == index && input.is_analog_input(i) {
+                    d_ticks += input.get_analog_diff_and_reset(i, 200);
                 }
             }
         }
@@ -93,7 +93,7 @@ impl MusicSelectKeyProperty {
         for i in 0..assign.len() {
             for &index in assign[i] {
                 if code == index && input.get_key_state(i as i32) {
-                    if input.is_analog_input(i as i32) {
+                    if input.is_analog_input(i) {
                         continue;
                     }
                     if reset_state {
